requirements.txt

django
pillow
django-crispy-forms
crispy-bootstrap5
django-rest-framework
coreapi